Book2 Chapter 1 - Motionblur

+++ RaytracerWeekend/vec3.h
@@ -52,7 +52,7 @@ inline std::ostream& operator>>(std::ostream &os, vec3 &t) {
 }
 
 inline void vec3::make_unit_vector() {
-	float k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+	float k =(float)1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
 	e[0] *= k;
 	e[1] *= k;
 	e[2] *= k;	
@@ -132,7 +132,7 @@ inline vec3& vec3::operator*=(const float t) {
 }
 
 inline vec3& vec3::operator/=(const float t) {
-	float k = 1.0/t;
+	float k = (float)1.0/t;
 	e[0] /= t;
 	e[1] /= t;
 	e[2] /= t;
